+++ modules/farm_rothamsted_quick/src/Plugin/QuickForm/QuickTrailerHarvest.php
@@ -320,6 +320,7 @@ class QuickTrailerHarvest extends QuickExperimentFormBase {
 
     // Compute the trailer nett weight.
     $total_nett_weight = [
+      'type' => 'standard',
       'label' => 'Nett weight',
       'measure' => 'weight',
       'units' => NULL,